9 TESTS DONE

+++ test07.sh
@@ -31,9 +31,6 @@ pigs-init> /dev/null 2>&1
 pigs-add a > /dev/null 2>&1
 pigs-commit -m 'test'> /dev/null 2>&1
 
-pigs-add b > /dev/null 2>&1
-pigs-commit -m 'second com' > /dev/null 2>&1
-
 # ERROR: Change in WORKING vs REPO AND INDEX
 echo line2 >> a
 
@@ -76,12 +73,45 @@ if ! diff "$expected_output" "$actual_output"; then
     exit 1
 fi
 
+
+cat > "$expected_output" <<EOF
+pigs-rm: error: 'a' in index is different to both the working file and the repository
+EOF
+
+pigs-rm --cached a> "$actual_output" 2>&1
+sed -i 's|^.*/||' "$actual_output"
+if ! diff "$expected_output" "$actual_output"; then
+    echo "Failed test"
+    exit 1
+fi
+
+# CRITICAL ERRORS (UN FORCE-ABLE)
 #  ERROR: Filename doesnt exist
 cat > "$expected_output" <<EOF
 pigs-rm: error: 'b' is not in the pigs repository
 EOF
 
 pigs-rm b> "$actual_output" 2>&1
+sed -i 's|^.*/||' "$actual_output"
+if ! diff "$expected_output" "$actual_output"; then
+    echo "Failed test"
+    exit 1
+fi
+
+# ERROR: File only exists in index
+pigs-rm --cached --force a > /dev/null 2>&1
+cat > "$expected_output" <<EOF
+pigs-rm: error: 'a' is not in the pigs repository
+EOF
+
+pigs-rm a> "$actual_output" 2>&1
+sed -i 's|^.*/||' "$actual_output"
+if ! diff "$expected_output" "$actual_output"; then
+    echo "Failed test"
+    exit 1
+fi
+
+pigs-rm --force a> "$actual_output" 2>&1
 sed -i 's|^.*/||' "$actual_output"
 if ! diff "$expected_output" "$actual_output"; then
     echo "Failed test"
